pyproject.toml
[project]
name = "spotify-playlists"
version = "0.1.0"
description = ""
authors = [{ name = "Ted Lawless", email = "lawlesst@gmail.com" }]
requires-python = ">3.11,<3.14"
readme = "README.md"
dependencies = [
    "python-dotenv>=1.0.0,<2",
    "requests>=2.31.0,<3",
    "bs4>=0.0.2,<0.0.3",
    "pydantic>=2.11.9,<3",
    "click>=8.3.0",
]

[dependency-groups]
dev = [
    "ipython>=8.14.0,<9",
    "ipdb>=0.13.13,<0.14",
    "pre-commit>=3.3.3,<4",
]
gemini = [
    "google-genai>=1.50.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["spotify", "scrapers"]

[tool.uv]
package = true
